refactor

+++ src/main/java/com/myapp/repositories/EstadoPedidoRepository.java
@@ -1,8 +1,6 @@
 package com.myapp.repositories;
 
 
-import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.myapp.models.EstadoPedido;
@@ -10,6 +8,5 @@ import com.myapp.models.EstadoPedido;
 
 
 @Repository
-public interface EstadoPedidoRepository extends JpaRepository<EstadoPedido, Long>{
-
+public interface EstadoPedidoRepository extends JpaRepository<EstadoPedido, Integer>{
 }
